Remove legacy renderers.
